[#110] 커뮤니티 게시판 상세 페이지 컴포넌트 분리  (#111)
* feat: 커뮤니티 상세 페이지 ssr/csr 분리 및 axios 인스턴스 수정

* feat: 불필요 react 텍스트 삭제

* fix: 추천 변수명 수정

* chore: 게시물 상세내용 마크다운으로 변경

* refactor: DOMPurify 보안 방식으로 변경

* refactor: 불필요 코드 수정

* chore: 커뮤니티 게시판 업로드 이미지 삭제

* chore: 파일명 수정

* feat: 커뮤니티 게시물 응답 에러 코드 추가

* chore: 커스텀 에러 파라미터 string으로 변경

* chore: 전체 커스텀 에러 메세지 string 변경
