+++ apps/next-client/package.json
@@ -18,7 +18,7 @@
     "bootstrap": "^5.3.3",
     "dompurify": "^3.2.3",
     "fast-xml-parser": "^4.5.0",
-    "isomorphic-dompurify": "^2.16.0",
+    "isomorphic-dompurify": "^2.20.0",
     "js-cookie": "^3.0.5",
     "jsdom": "^25.0.1",
     "next": "^14.2.13",
@@ -27,8 +27,11 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
+    "react-markdown": "^9.0.3",
     "react-quill": "^2.0.0",
     "react-toastify": "^11.0.2",
+    "rehype-raw": "^7.0.0",
+    "sanitize-html": "^2.14.0",
     "sass": "^1.79.4",
     "styled-components": "^6.1.13",
     "zustand": "^5.0.3"
@@ -40,6 +43,7 @@
     "@types/node": "^20.14.9",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
+    "@types/sanitize-html": "^2.13.0",
     "eslint": "^8.57.0",
     "eslint-config-next": "14.2.4",
     "eslint-config-prettier": "^9.1.0",
