+++ apps/next-client/src/components/community/ForwardedReactQuill.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import React, { forwardRef, useRef, useImperativeHandle } from 'react';
+import { forwardRef, useRef, useImperativeHandle } from 'react';
 import ReactQuill, { ReactQuillProps } from 'react-quill';
 import type Quill from 'quill';
 import { ERROR_MESSAGES } from '@/constants/errors';
