+++ apps/next-client/src/error/AuthError.ts
@@ -1,14 +1,14 @@
 import { ERROR_MESSAGES } from '@/constants/errors';
 
 export class SignupError extends Error {
-  constructor(message = ERROR_MESSAGES.SIGN_UP_ERROR) {
+  constructor(message: string = ERROR_MESSAGES.SIGN_UP_ERROR) {
     super(message);
     this.name = "SignupError";
   }
 }
 
 export class LoginError extends Error {
-  constructor(message = ERROR_MESSAGES.LOGIN_ERROR) {
+  constructor(message: string = ERROR_MESSAGES.LOGIN_ERROR) {
     super(message);
     this.name = "LoginError";
   }
