+++ apps/next-client/src/error/CredError.ts
@@ -1,7 +1,7 @@
 import { ERROR_MESSAGES } from '@/constants/errors';
 
 export class CredError extends Error {
-  constructor(message = ERROR_MESSAGES.INVALID_CREDENTIAL) {
+  constructor(message: string = ERROR_MESSAGES.INVALID_CREDENTIAL) {
     super(message);
     this.name = "CredError";
   }
