+++ apps/next-client/src/error/MypageError.ts
@@ -1,28 +1,28 @@
-import { ALERT_MESSAGES } from '@/constants/alert_message';
+import { ALERT_MESSAGES } from '@/constants/alertMessage';
 
 export class FetchUsernameError extends Error {
-  constructor(message = ALERT_MESSAGES.ERROR.FETCH_USERNAME) {
+  constructor(message: string = ALERT_MESSAGES.ERROR.FETCH_USERNAME) {
     super(message);
     this.name = "FetchUsernameError";
   }
 }
 
 export class UpdateNicknameError extends Error {
-  constructor(message = ALERT_MESSAGES.ERROR.UPDATE_NICKNAME) {
+  constructor(message: string = ALERT_MESSAGES.ERROR.UPDATE_NICKNAME) {
     super(message);
     this.name = "UpdateNicknameError";
   }
 }
 
 export class UpdatePasswordError extends Error {
-  constructor(message = ALERT_MESSAGES.ERROR.UPDATE_PASSWORD) {
+  constructor(message: string = ALERT_MESSAGES.ERROR.UPDATE_PASSWORD) {
     super(message);
     this.name = "UpdatePasswordError";
   }
 }
 
 export class DeleteAccountError extends Error {
-  constructor(message = ALERT_MESSAGES.ERROR.DELETE_ACCOUNT) {
+  constructor(message: string = ALERT_MESSAGES.ERROR.DELETE_ACCOUNT) {
     super(message);
     this.name = "DeleteAccountError";
   }
