+++ apps/next-client/src/error/SearchError.ts
@@ -2,7 +2,7 @@ import { SEARCH_ERROR_MESSAGES } from '@/constants/search_errors';
 
 // API 키 누락 에러
 export class ApiKeyMissingError extends Error {
-  constructor(message = SEARCH_ERROR_MESSAGES.API_KEY_MISSING) {
+  constructor(message: string = SEARCH_ERROR_MESSAGES.API_KEY_MISSING) {
     super(message);
     this.name = 'ApiKeyMissingError';
   }
@@ -10,7 +10,7 @@ export class ApiKeyMissingError extends Error {
 
 // API 요청 실패 에러
 export class ApiRequestError extends Error {
-  constructor(message = SEARCH_ERROR_MESSAGES.API_REQUEST_ERROR) {
+  constructor(message: string = SEARCH_ERROR_MESSAGES.API_REQUEST_ERROR) {
     super(message);
     this.name = 'ApiRequestError';
   }
@@ -18,7 +18,7 @@ export class ApiRequestError extends Error {
 
 // API 응답 파싱 에러
 export class ApiResponseParsingError extends Error {
-  constructor(message = SEARCH_ERROR_MESSAGES.API_RESPONSE_PARSING_ERROR) {
+  constructor(message: string = SEARCH_ERROR_MESSAGES.API_RESPONSE_PARSING_ERROR) {
     super(message);
     this.name = 'ApiResponseParsingError';
   }
@@ -26,7 +26,7 @@ export class ApiResponseParsingError extends Error {
 
 // 검색어가 너무 짧을 때 에러
 export class ShortSearchTermError extends Error {
-  constructor(message = SEARCH_ERROR_MESSAGES.SHORT_SEARCH_TERM) {
+  constructor(message: string = SEARCH_ERROR_MESSAGES.SHORT_SEARCH_TERM) {
     super(message);
     this.name = 'ShortSearchTermError';
   }
@@ -34,7 +34,7 @@ export class ShortSearchTermError extends Error {
 
 // 검색 결과가 없을 때 에러
 export class NoResultsError extends Error {
-  constructor(message = SEARCH_ERROR_MESSAGES.NO_RESULTS_FOUND) {
+  constructor(message: string = SEARCH_ERROR_MESSAGES.NO_RESULTS_FOUND) {
     super(message);
     this.name = 'NoResultsError';
   }
