+++ apps/next-client/src/utils/authUtils.ts
@@ -1,7 +1,7 @@
 import { axiosInstance } from '@/services/axiosInstance';
 import { API_URLS } from '@/constants/urls';
 import { JWT } from 'next-auth/jwt';
-import { ALERT_MESSAGES } from "@/constants/alert_message";
+import { ALERT_MESSAGES } from "@/constants/alertMessage";
 import Cookies from "js-cookie";
 
 const ACCESS_TOKEN_EXPIRES_IN = 60 * 60 * 1000;
